+++ services/common/src/@Types/declarations.d.ts
@@ -7,13 +7,3 @@ declare module "*.mp3" {
   const value: any;
   export default value;
 }
-
-declare module "*.gif" {
-  const value: any;
-  export default value;
-}
-
-declare module "*.json" {
-  const value: any;
-  export default value;
-}
